Move token parsers & nodes to namespace
